+++ src/main/java/org/folio/marc4ld/service/marc2ld/authority/identifier/Marc010IdentifierMapper.java
@@ -1,0 +1,42 @@
+package org.folio.marc4ld.service.marc2ld.authority.identifier;
+
+import static org.folio.marc4ld.util.Constants.A;
+import static org.folio.marc4ld.util.Constants.TAG_010;
+import static org.folio.marc4ld.util.MarcUtil.getSubfieldValueWithoutSpaces;
+
+import java.util.Optional;
+import org.folio.ld.dictionary.model.Resource;
+import org.folio.ld.fingerprint.service.FingerprintHashService;
+import org.folio.marc4ld.service.label.LabelService;
+import org.folio.marc4ld.service.marc2ld.mapper.MapperHelper;
+import org.marc4j.marc.Record;
+import org.springframework.stereotype.Component;
+
+@Component
+public class Marc010IdentifierMapper extends AbstractIdentifierMapper {
+
+  public Marc010IdentifierMapper(LabelService labelService,
+                                 IdentifierLinkService identifierLinkService,
+                                 MapperHelper mapperHelper,
+                                 FingerprintHashService hashService) {
+    super(labelService, identifierLinkService, mapperHelper, hashService);
+  }
+
+  @Override
+  public boolean isApplicable(Record marcRecord) {
+    return get010SubfieldA(marcRecord).isPresent();
+  }
+
+  @Override
+  public void map(Record marcRecord, Resource authority) {
+    get010SubfieldA(marcRecord)
+      .ifPresent(id -> addIdentifierEdge(authority, id));
+  }
+
+  private Optional<String> get010SubfieldA(Record marc) {
+    return marc.getDataFields().stream()
+      .filter(df -> df.getTag().equals(TAG_010))
+      .flatMap(df -> getSubfieldValueWithoutSpaces(df, A).stream())
+      .findFirst();
+  }
+}