+++ src/main/java/org/folio/marc4ld/util/Constants.java
@@ -38,6 +38,7 @@ public class Constants {
   public static final char Z = 'z';
   public static final String TAG_005 = "005";
   public static final String TAG_008 = "008";
+  public static final String TAG_010 = "010";
   public static final String TAG_024 = "024";
   public static final String TAG_043 = "043";
   public static final String TAG_100 = "100";