+++ src/test/java/org/folio/marc4ld/authority/field111/MarcToLdAuthorityConceptMeeting111IT.java
@@ -2,6 +2,9 @@ package org.folio.marc4ld.authority.field111;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.CONCEPT;
+import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_FAST;
+import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_LCNAF;
+import static org.folio.ld.dictionary.ResourceTypeDictionary.ID_LOCAL;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.MEETING;
 import static org.folio.marc4ld.mapper.test.TestUtil.loadResourceAsString;
 import static org.folio.marc4ld.mapper.test.TestUtil.validateResource;
@@ -39,7 +42,7 @@ class MarcToLdAuthorityConceptMeeting111IT extends Marc2LdTestBase {
         resource -> validateResource(resource, List.of(CONCEPT, MEETING), generalProperties(), EXPECTED_MAIN_LABEL))
       .satisfies(resource -> validateFocusResource(resource, MEETING, focusProperties(), EXPECTED_FOCUS_LABEL))
       .satisfies(AuthorityValidationHelper::validateSubFocusResources)
-      .satisfies(resource -> validateIdentifier(resource, "010fieldvalue"));
+      .satisfies(resource -> validateIdentifier(resource, "85121033", ID_LOCAL, null));
   }
 
   @Test
@@ -53,9 +56,10 @@ class MarcToLdAuthorityConceptMeeting111IT extends Marc2LdTestBase {
     //then
     assertThat(resources)
       .singleElement()
-      .satisfies(resource -> assertThat(resource.getOutgoingEdges()).hasSize(1))
+      .satisfies(resource -> assertThat(resource.getOutgoingEdges()).hasSize(2))
       .satisfies(resource -> validateResource(resource, List.of(MEETING), meetingProperties(), EXPECTED_MEETING_LABEL))
-      .satisfies(resource -> validateIdentifier(resource, "010fieldvalue"));
+      .satisfies(resource -> validateIdentifier(resource, "fst12345678", ID_FAST, null))
+      .satisfies(resource -> validateIdentifier(resource, "n0987654321", ID_LCNAF, "http://id.loc.gov/authorities/n0987654321"));
   }
 
   @ParameterizedTest